requirements.txt
arrow==1.2.3
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
ndjson==0.3.1
pick==2.0.2
psycopg==3.1.4
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
windows-curses==2.3.0